Extract QR code generation code to separate files
Which is really platform-dependent, but still can't go into os/
because they need wxWidgets (for creating a wxBitmap)
