+++ src/ui/wxWidgets/CMakeLists.txt
@@ -111,8 +111,8 @@ if(HAVE_YKPERS_H)
 endif(HAVE_YKPERS_H)
 
 if (NOT NO_QR)
-  list(APPEND PWSAFE_SRC_LIST src/ui/wxWidgets/pwsqrcodedlg.cpp)
-  list(APPEND PWSAFE_HDR_LIST src/ui/wxWidgets/yubicfg.h src/ui/wxWidgets/pwsqrcodedlg.h)
+  list(APPEND PWSAFE_SRC_LIST src/ui/wxWidgets/pwsqrcodedlg.cpp src/ui/wxWidgets/pwsqrencode-linux.cpp)
+  list(APPEND PWSAFE_HDR_LIST src/ui/wxWidgets/pwsqrencode.h src/ui/wxWidgets/pwsqrcodedlg.h)
 endif(NOT NO_QR)
 
 set (PWSAFE_SRCS ${PWSAFE_SRC_LIST} PARENT_SCOPE)
